requirements.txt
Pillow==8.0.1
google-cloud-vision==2.0.0
